src/commands/ralsei/ralsei-sprite/sprites/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.3.0
requests==2.32.4
tqdm==4.67.1
urllib3==2.5.0
